# tax connect/services/payment-platform/openapi.yaml
openapi: 3.1.0
info:
  title: Payment and Agreement Platform
  version: 0.1.0
servers:
  - url: https://api.taxconnect.local
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /negotiations:
    post:
      summary: Create negotiation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NegotiationCreate"
      responses:
        "201":
          description: Created
  /negotiations/{id}:
    get:
      summary: Get negotiation and proposals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /negotiations/{id}/proposals:
    post:
      summary: Submit proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProposalCreate"
      responses:
        "201":
          description: Created
  /negotiations/{id}/accept-final:
    post:
      summary: Accept final offer and create agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
  /agreements/{id}:
    get:
      summary: Get agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /agreements/{id}/sign:
    post:
      summary: Upload signed agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Signed
  /agreements/{id}/payments/advance:
    post:
      summary: Initiate advance payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "201":
          description: Created
        "402":
          description: Escrow conditions not met
  /agreements/{id}/payments/balance:
    post:
      summary: Initiate balance payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "201":
          description: Created
        "402":
          description: Escrow conditions not met
  /agreements/{id}/payments/{paymentId}/confirm:
    post:
      summary: Confirm payment success
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /agreements/{id}/escrow/release:
    post:
      summary: Release escrow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /agreements/{id}/disputes:
    post:
      summary: Raise dispute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeCreate"
      responses:
        "201":
          description: Created
    get:
      summary: List disputes for agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /audit:
    get:
      summary: Search audit logs
      parameters:
        - name: agreementId
          in: query
          schema:
            type: string
        - name: actorId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
components:
  schemas:
    NegotiationCreate:
      type: object
      required: [clientId, caId, scopeText, deliverables, acceptanceCriteria]
      properties:
        clientId:
          type: string
        caId:
          type: string
        scopeText:
          type: string
        deliverables:
          type: array
          items:
            type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
    ProposalCreate:
      type: object
      required: [amountInr, scopeText, deliverables, acceptanceCriteria, proposalType]
      properties:
        amountInr:
          type: number
        scopeText:
          type: string
        deliverables:
          type: array
          items:
            type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        proposalType:
          type: string
          enum: [PROPOSAL, COUNTER, FINAL]
    PaymentRequest:
      type: object
      required: [amountInr, gstinCa, hsnSac, paymentMethod]
      properties:
        amountInr:
          type: number
        gstRate:
          type: number
        gstType:
          type: string
          enum: [CGST_SGST, IGST]
        gstinClient:
          type: string
        gstinCa:
          type: string
        hsnSac:
          type: string
        paymentMethod:
          type: string
          enum: [UPI, NETBANKING, NEFT_RTGS, CARD]
        enforce3ds:
          type: boolean
    DisputeCreate:
      type: object
      required: [reason]
      properties:
        reason:
          type: string